requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiogram-dialog==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
asyncio==3.4.3
asyncpg==0.29.0
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
frozenlist==1.4.1
greenlet==3.1.0
idna==3.8
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.1.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
redis==5.0.8
SQLAlchemy==2.0.34
typing_extensions==4.12.2
yarl==1.11.1
